добавлен RV для отображения фото которые добавляются в публикацию

+++ app/src/main/java/com/myblogtour/blogtour/ui/addPublication/AddPublicationViewModel.kt
@@ -8,6 +8,7 @@ import androidx.lifecycle.ViewModel
 import com.google.firebase.storage.StorageReference
 import com.google.gson.JsonArray
 import com.google.gson.JsonObject
+import com.myblogtour.blogtour.domain.ImagePublicationEntity
 import com.myblogtour.blogtour.domain.repository.AuthFirebaseRepository
 import com.myblogtour.blogtour.domain.repository.CreatePublicationRepository
 import com.myblogtour.blogtour.domain.repository.ImageFbRepository
@@ -25,12 +26,12 @@ class AddPublicationViewModel(
     private var flagAddPublication = true
     private var counterImagePublication = 0
     private var loadingImagePublication = true
+    private lateinit var imagePublicationEntity: ImagePublicationEntity
+    private var listImagePublication: MutableList<ImagePublicationEntity> = mutableListOf()
 
     override val publishPostLiveData: LiveData<Boolean> = MutableLiveData()
-    override val loadUriOneImage: LiveData<Uri?> = MutableLiveData()
-    override val loadUriTwoImage: LiveData<Uri?> = MutableLiveData()
-    override val loadUriThreeImage: LiveData<Uri?> = MutableLiveData()
-    override val progressLoad: LiveData<Int> = MutableLiveData()
+    override val loadUriImage: LiveData<ImagePublicationEntity> = MutableLiveData()
+    override val progressLoad: LiveData<ImagePublicationEntity> = MutableLiveData()
     override val errorMessageImage: LiveData<String> = MutableLiveData()
     override val errorMessageText: LiveData<String> = MutableLiveData()
     override val errorMessageLocation: LiveData<String> = MutableLiveData()
@@ -38,26 +39,41 @@ class AddPublicationViewModel(
     override val address: LiveData<Editable> = SingleLiveEvent()
     override val errorAddress: LiveData<String> = SingleLiveEvent()
     override val counterImage: LiveData<Boolean> = SingleLiveEvent()
+    override val loadingImage: LiveData<Boolean> = SingleLiveEvent()
 
     fun flagAddPublication(b: Boolean) {
         flagAddPublication = b
     }
 
-    fun counterImage() {
+    fun getCounterImage() {
         if (counterImagePublication == 3) {
             counterImage.mutable().postValue(false)
         }
         return counterImage.mutable().postValue(true)
     }
 
-    fun deleteImage(uriImageOne: Uri?, uriImageTwo: Uri?, uriImageThree: Uri?) {
+    fun getLoadingImage() {
+        loadingImage.mutable().postValue(loadingImagePublication)
+    }
+
+    fun deleteImage() {
         val uriList: MutableList<Uri?> = mutableListOf()
-        with(uriList) {
-            add(uriImageOne)
-            add(uriImageTwo)
-            add(uriImageThree)
+        for (i in listImagePublication.indices) {
+            when (i) {
+                0 -> {
+                    uriList.add(listImagePublication[i].uriLocal)
+                }
+                1 -> {
+                    uriList.add(listImagePublication[i].uriLocal)
+                }
+                2 -> {
+                    uriList.add(listImagePublication[i].uriLocal)
+                }
+            }
         }
-        imageFbRepository.deleteImage(uriList.toList())
+        if (uriList.size > 0) {
+            imageFbRepository.deleteImage(uriList.toList())
+        }
     }
 
     fun deleteImage(uriImage: Uri?) {
@@ -66,37 +82,28 @@ class AddPublicationViewModel(
         }
     }
 
-    fun image(uri: Uri, imageNumber: Int) {
-        imageFbRepository.imageLoading(uri, onSuccess = {
-            when (imageNumber) {
-                1 -> {
-                    loadUriOneImage.mutable().postValue(it)
-                }
-                2 -> {
-                    loadUriTwoImage.mutable().postValue(it)
-                }
-                3 -> {
-                    loadUriThreeImage.mutable().postValue(it)
-                }
-            }
-
+    fun image(uri: Uri) {
+        loadingImagePublication = false
+        imageFbRepository.imageLoading(uri, onSuccess = { uriFb ->
+            imagePublicationEntity = ImagePublicationEntity(uriFb, uri, false)
+            listImagePublication.add(imagePublicationEntity)
+            loadUriImage.mutable().postValue(imagePublicationEntity)
+            loadingImagePublication = true
         }, onError = {
 
-        }, onProgress = {
-            progressLoad.mutable().postValue(it)
+        }, onProgress = { onProgress ->
+            imagePublicationEntity = ImagePublicationEntity(uriLocal = uri, progress = onProgress)
+            progressLoad.mutable().postValue(imagePublicationEntity)
         })
     }
 
     override fun dataPublication(
         text: String,
         location: String,
-        uriImageOne: Uri?,
-        uriImageTwo: Uri?,
-        uriImageThree: Uri?,
     ) {
         if (flagAddPublication) {
             when {
-                imageUriNotNull(uriImageOne, uriImageTwo, uriImageThree) == null -> {
+                listImagePublication.size == 0 -> {
                     errorMessageImage.mutable().postValue("Добавьте изображение")
                     return
                 }
@@ -112,9 +119,6 @@ class AddPublicationViewModel(
                     flagAddPublication = false
                     val publishPost =
                         converterJsonObject(
-                            uriImageOne,
-                            uriImageTwo,
-                            uriImageThree,
                             text,
                             location)
                     createPublicationRepository.createPublication(onSuccess = {
@@ -135,21 +139,7 @@ class AddPublicationViewModel(
         })
     }
 
-    private fun imageUriNotNull(
-        uriImageOne: Uri?,
-        uriImageTwo: Uri?,
-        uriImageThree: Uri?,
-    ): Uri? {
-        if (uriImageOne != null || uriImageTwo != null || uriImageThree != null) {
-            return Uri.parse("")
-        }
-        return null
-    }
-
     private fun converterJsonObject(
-        uriImageOne: Uri?,
-        uriImageTwo: Uri?,
-        uriImageThree: Uri?,
         text: String,
         location: String,
     ): JsonObject {
@@ -162,33 +152,41 @@ class AddPublicationViewModel(
         }, onError = {
 
         })
-        val urlImage = JsonArray()
-        val image = JsonObject()
-        val imageTwo = JsonObject()
-        val imageThree = JsonObject()
         val publicationJson = JsonObject()
         val fieldsJson = JsonObject()
-        uriImageOne?.let {
-            image.addProperty("url", it.toString())
-            urlImage.add(image)
-        }
-        uriImageTwo?.let {
-            imageTwo.addProperty("url", it.toString())
-            urlImage.add(imageTwo)
-        }
-        uriImageThree?.let {
-            imageThree.addProperty("url", it.toString())
-            urlImage.add(imageThree)
-        }
         userProfile.add(userIdProfile)
         with(publicationJson) {
             addProperty("text", text)
             addProperty("location", location)
-            add("image", urlImage)
+            add("image", converterFromListToJson())
             add("userprofile", userProfile)
         }
         fieldsJson.add("fields", publicationJson)
         return fieldsJson
+    }
+
+    private fun converterFromListToJson(): JsonArray {
+        val urlImage = JsonArray()
+        val image = JsonObject()
+        val imageTwo = JsonObject()
+        val imageThree = JsonObject()
+        for (i in listImagePublication.indices) {
+            when (i) {
+                0 -> {
+                    image.addProperty("url", listImagePublication[i].url.toString())
+                    urlImage.add(image)
+                }
+                1 -> {
+                    imageTwo.addProperty("url", listImagePublication[i].url.toString())
+                    urlImage.add(imageTwo)
+                }
+                2 -> {
+                    imageThree.addProperty("url", listImagePublication[i].url.toString())
+                    urlImage.add(imageThree)
+                }
+            }
+        }
+        return urlImage
     }
 
     private fun <T> LiveData<T>.mutable(): MutableLiveData<T> {
